build changes, working on quest logic
