+++ frontend/src/components/smart/QuestReward.vue
@@ -12,7 +12,7 @@
 <script lang="ts">
 import Vue from 'vue';
 import {PropType} from 'vue/types/options';
-import {QuestItemType, Rarity} from '@/views/Quests.vue';
+import { QuestItemType, Rarity } from '@/interfaces';
 import NftList, {NftIdType} from '@/components/smart/NftList.vue';
 import QuestComponentIcon from '@/components/smart/QuestComponentIcon.vue';
 import {questItemTypeSupportsStars, questItemTypeSupportsTimesValue} from '../../utils/common';
