+++ frontend/src/components/smart/QuestRow.vue
@@ -7,7 +7,7 @@
     <QuestRequirements v-if="character.quest && character.quest.id !== 0" :quest="character.quest"
                        :progress="character.quest.progress" :index="characterId"/>
     <QuestRewards v-if="character.quest && character.quest.id !== 0" :quest="character.quest"/>
-    <QuestActions :character="character" :quest="character.quest" :key="character.quest.id" showSupply
+    <QuestActions :character="character" :quest="character.quest" :pickable="pickable" :key="character.quest.id" showSupply
                   @refresh-quest-data="onRefreshQuestData" :questTemplateType="questTemplateType"/>
   </div>
   <div v-if="questTemplateType === QuestTemplateType.WALLET && quest" class="quest-row-wallet">
@@ -26,7 +26,12 @@ import QuestCharacter from '@/components/smart/QuestCharacter.vue';
 import QuestRequirements from '@/components/smart/QuestRequirements.vue';
 import QuestRewards from '@/components/smart/QuestRewards.vue';
 import QuestActions from '@/components/smart/QuestActions.vue';
-import {Quest, ReputationLevelRequirements, Rarity, RewardType, QuestTemplateType} from '@/views/Quests.vue';
+import {
+  Quest,
+  ReputationLevelRequirements,
+  Rarity,
+  RewardType,
+  QuestTemplateType } from '@/interfaces';
 import {mapActions, mapGetters} from 'vuex';
 import {Nft, NftStatus} from '@/interfaces/Nft';
 
@@ -59,6 +64,10 @@ export default Vue.extend({
     },
     questTemplateType: {
       type: Number as PropType<QuestTemplateType>,
+    },
+    pickable: {
+      type: Boolean,
+      default: false
     },
   },
