+++ frontend/src/interfaces/index.ts
@@ -4,3 +4,4 @@ export * from './State';
 export * from './Target';
 export * from './Weapon';
 export * from './Contracts';
+export * from './Quest';
